LU-13697 llite: fix short io for AIO
The problem is currently AIO could not handle i/o size > stripe size:

We need cl io loop to handle io across stripes, since -EIOCBQUEUED is
returned for AIO, io loop will be stopped thus short io happen.

The patch try to fix the problem by making IO engine aware of
special error, and it will be proceed to finish all IO requests.

Fixes: d1dde ("LU-4198 clio: AIO support for direct IO")
Reviewed-on: https://review.whamcloud.com/39104
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>

+++ lustre/tests/sanity.sh
@@ -21828,6 +21828,13 @@ test_398c() { # LU-4198
 		--filename=$DIR/$tfile
 	[ $? -eq 0 ] || error "fio mixed read write error"
 
+	echo "AIO with large block size ${size}M"
+	fio --name=rand-rw --rw=randrw --bs=${size}M --direct=1 \
+		--numjobs=1 --fallocate=none --ioengine=libaio \
+		--iodepth=16 --allow_file_create=0 --size=${size}M \
+		--filename=$DIR/$tfile
+	[ $? -eq 0 ] || error "fio large block size failed"
+
 	rm -rf $DIR/$tfile
 	$LCTL set_param debug="$saved_debug"
 }
